+++ scripts/macros/2024/spells/magicMissile.js
@@ -2,5 +2,16 @@ import {magicMissile as magicMissileLegacy} from '../../../legacyMacros.js';
 export let magicMissile = {
     ...magicMissileLegacy,
     version: '1.2.22',
-    rules: 'modern'
+    rules: 'modern',
+    config: [
+        {
+            value: 'rollEach',
+            label: 'CHRISPREMADES.Macros.MagicMissile.RollEach',
+            type: 'checkbox',
+            default: true,
+            category: 'homebrew',
+            homebrew: true
+        },
+        ...magicMissileLegacy.config.slice(1)
+    ]
 };