+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.2.25 Release:
+## Bug Fixes:
+- Magic Missile (Legacy & Modern)
+- Eldritch Blast (Legacy & Modern)
+- Automated Animations integration
+
 # 1.2.24 Release:
 ## New Modern Content:
 - True Strike